:recycle: Extract ModVersion.ts
